updates: 
* removing the difference in the arg_type between CXX11 and not CXX11 (::type postfix)
* adding unit tests for the different interfaces
few other changes

Conflicts:
	include/stencil-composition/backend.h
	include/stencil-composition/backend_cuda/backend_cuda.h
	include/stencil-composition/backend_host/backend_host.h
	include/stencil-composition/basic_token_execution.h
	include/stencil-composition/execution_policy.h
	include/stencil-composition/heap_allocated_temps.h
	include/stencil-composition/iterate_domain.h

+++ include/common/defs.h
@@ -84,10 +84,10 @@ namespace gridtools{  namespace enumtype{
 
 #include <boost/mpl/integral_c.hpp>
 #ifdef CXX11_ENABLED
-using int_t          = int;
+using int_t          = long int;
 using short_t        = int;
-using uint_t         = int;
-using ushort_t       = int;
+using uint_t         = unsigned long int;
+using ushort_t       = unsigned int;
 template<int_t N>
 using  static_int=boost::mpl::integral_c<int_t,N>;
 template<uint_t N>
@@ -97,10 +97,10 @@ using  static_short=boost::mpl::integral_c<short_t,N>;
 template<ushort_t N>
 using  static_ushort=boost::mpl::integral_c<ushort_t,N>;
 #else
-typedef int                     int_t;
+typedef long int                     int_t;
 typedef int                     short_t;
-typedef int            uint_t;
-typedef int            ushort_t;
+typedef unsigned long int            uint_t;
+typedef unsigned int            ushort_t;
 template<int_t N>
 struct static_int : boost::mpl::integral_c<int_t,N>{
     typedef boost::mpl::integral_c<int_t,N> type;
